requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
numpy==1.14.5
packaging==16.8
protobuf==3.6.0
pyparsing==2.1.10
six==1.10.0
tensorflow
Werkzeug==0.11.15
Keras==2.1.6
Pillow==4.2.1
opencv-python
pandas
scikit-learn==0.19.1
scikit-image==0.13.0